Ser kva som skjer om det blir ein r&r-app
